Improve colors
The original set was missing blue (thanks, David), but it's a good time
to choose nicer colors.

The default CSS color words like "blue" or "yellow" were chosen back in
the era of 16-bit colors, and the colors chosen were simply the corners
of the color space, chosen for mathematical simplicity.

Actual colors people want to draw with are generally less harsh, and
easier to see on a white background, which is what these colors
provide.